+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased](https://github.com/contentful/contentful.php/compare/2.0.0...HEAD)
+
 ## [2.0.0](https://github.com/contentful/contentful.php/tree/2.0.0) (2017-06-13)
 
 **ATTENTION**: This release contains breaking changes. Please take extra care when updating to this version.